requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
fonttools==4.37.4
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
